refactor(slides): separate composition and share text truncation
