+++ packages/core/src/slides/text-transcript.ts
@@ -38,19 +38,16 @@ export function parseTimestampSeconds(value: string): number | null {
     if (seconds >= 60) return null;
     return minutes * 60 + seconds;
   }
-  if (parts.length === 3) {
-    const [hours, minutes, seconds] = parts;
-    if (minutes >= 60 || seconds >= 60) return null;
-    return hours * 3600 + minutes * 60 + seconds;
-  }
-  return null;
+  const [hours, minutes, seconds] = parts;
+  if (minutes >= 60 || seconds >= 60) return null;
+  return hours * 3600 + minutes * 60 + seconds;
 }
 
 function normalizeSlideText(value: string): string {
   return value.replace(/\s+/g, " ").trim();
 }
 
-function truncateSlideText(value: string, limit: number): string {
+export function truncateSlideText(value: string, limit: number): string {
   if (value.length <= limit) return value;
   const truncated = value.slice(0, limit).trimEnd();
   const clean = truncated.replace(/\s+\S*$/, "").trim();
